pyproject.toml
[project]
name = "backend-curavita"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.3",
    "celery-types>=0.26.0",
    "django>=6.0.4",
    "django-allauth>=65.15.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-phonenumber-field>=8.4.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.9",
    "drf-spectacular>=0.29.0",
    "geopy>=2.4.1",
    "phonenumbers>=9.0.27",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
]
